DTA: Implement support for BufferID metadata.

+++ test/simple-test-cases/bufferid.c
@@ -4,12 +4,12 @@
 
 int main(int argc, char *argv[])
 {
-  __attribute__((annotate("scalar(range(-100, 100)) bufferid('buf1')"))) int buffer1[100];
-  __attribute__((annotate("scalar(range(-10000, 10000)) bufferid('buf1')"))) int buffer2[100];
+  __attribute__((annotate("scalar(range(-100, 100)) bufferid('buf1')"))) float buffer1[100];
+  __attribute__((annotate("scalar(range(-10000, 10000)) bufferid('buf1')"))) float buffer2[100];
 
   for (int i=0; i<100; i++) {
-    int *pbuf1 = buffer1+i;
-    int *pbuf2 = &(buffer2[i]);
+    float *pbuf1 = buffer1+i;
+    float *pbuf2 = &(buffer2[i]);
     *pbuf1 = (float)i / 100;
     *pbuf2 = (float)i / 100;
   }
@@ -19,7 +19,7 @@ int main(int argc, char *argv[])
     if (buffer1[i] == buffer2[i])
       n++;
   }
-  printf("%d\n", n);
+  printf("%d\n", n); // prints 4 without bufferid, 100 with bufferid
 
   return 0;
 }
